refactor ResourceUsageLog
